feat: email send

+++ httpWork/ioTxt.js
@@ -2,17 +2,28 @@ const fs = require("fs");
 
 const path = require("path");
 
-const inputHandle = (content) => {
+const isExists = (filePath) => {
+  return new Promise((resolve, reject) => {
+    fs.access(filePath, fs.constants.F_OK, (err) => {
+      // if (err) console.log(`${file} ${err ? "does not exist" : "exists"}`);s
+      if (!err) {
+        resolve(true);
+      } else {
+        reject(false);
+      }
+    });
+  });
+};
+
+let n = 1;
+const inputHandle = async (content) => {
   // const content = fs.readFileSync("./src/template.jsx");
   //   let content = "teset22d\n";
 
-  let info = {};
-
-  let n = 1;
-
-  fs.rmdirSync("./log", { recursive: true, force: true });
-
-  fs.mkdirSync("./log");
+  if ((n = 1)) {
+    fs.rmdirSync("./log", { recursive: true, force: true });
+    fs.mkdirSync("./log");
+  }
 
   //   info[n] = "testcontent1";
 
@@ -23,8 +34,16 @@ const inputHandle = (content) => {
 
   //     content += info[key] + "\n";
   //   }
-
-  const filePath = path.resolve(__dirname, `log/log.json`);
+  let fileName = `log-${n}.json`;
+  let is = await isExists(path.resolve(__dirname, `log/${fileName}`));
+  console.log(is);
+  if (is) {
+    n = n + 1;
+    fileName = `log-${n}.json`;
+    console.log(n);
+  }
+  // const filePath = path.resolve(__dirname, `log/log.json`);
+  const filePath = path.resolve(__dirname, `log/${fileName}`);
   // const filePath2 = path.resolve(__dirname, `log/log.txt`);
   console.log(filePath);
   console.log("日志信息已记录");
